+++ src/parser/lexer.h
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "base/check.h"
-#include "parser/flex_input.h"
+#include "parser/input.h"
 #include "parser/tokens.h"
 
 namespace cvc5 {
@@ -50,23 +50,20 @@ std::ostream& operator<<(std::ostream& o, const Span& l);
 #define INPUT_BUFFER_SIZE 32768
 
 /**
- * A Flex lexer. This class inherits from yyFlexLexer, which is generated
- * by Flex's C++ code generation.
- *
- * Custom lexers (e.g. for smt2) override the yylex method of the base
- * class.
+ * The base lexer class. The main methods to override are initialize and
+ * nextTokenInternal.
  */
-class FlexLexer
+class Lexer
 {
  public:
-  FlexLexer();
-  virtual ~FlexLexer() {}
+  Lexer();
+  virtual ~Lexer() {}
   /**
    * Initialize the lexer to generate tokens from stream input.
    * @param input The input stream
    * @param inputName The name for debugging
    */
-  virtual void initialize(FlexInput* input, const std::string& inputName);
+  virtual void initialize(Input* input, const std::string& inputName);
   /**
    * String corresponding to the last token (old top of stack). This is only
    * valid if no tokens are currently peeked.