+++ NEWS.md
@@ -24,6 +24,9 @@ This file contains a summary of important user-visible changes.
        `Term::getRealAlgebraicNumberUpperBound()` may now be used to query
        the contents of terms corresponding to real algebraic numbers.
 - Removed support for the ANTLR parser and parsing for the TPTP language.
+- API: Removed API support for the deprecated SyGuS 2.0 command
+       `Solver::synthInv`. This method is equivalent to `Solver::synthFun`
+       with a Boolean range sort.
 
 cvc5 1.0.5
 ==========